[STYLE] 코드클리닝
